+++ feature/home/src/main/java/com/arifwidayana/home/domain/home/BannerUseCase.kt
@@ -1,9 +1,9 @@
-package com.arifwidayana.home.domain
+package com.arifwidayana.home.domain.home
 
 import com.arifwidayana.core.base.BaseUseCase
 import com.arifwidayana.core.wrapper.ViewResource
 import com.arifwidayana.home.data.network.repository.HomeRepository
-import com.arifwidayana.shared.data.network.model.mapper.home.BannerListMapper
+import com.arifwidayana.shared.data.network.model.mapper.home.BannerMapper
 import com.arifwidayana.shared.data.network.model.response.home.banner.BannerParamResponse
 import com.arifwidayana.shared.utils.ext.suspendSource
 import kotlinx.coroutines.CoroutineDispatcher
@@ -18,13 +18,13 @@ class BannerUseCase(
 ) : BaseUseCase<Nothing, BannerDataResource>(coroutineDispatcher) {
     override suspend fun execute(param: Nothing?): Flow<ViewResource<BannerDataResource>> = flow {
         emit(ViewResource.Loading())
-        homeRepository.showBanner().collect { source ->
-            source.suspendSource(
-                doOnSuccess = {
-                    emit(ViewResource.Success(BannerListMapper.toViewParam(it.payload)))
+        homeRepository.showBanner().collect {
+            it.suspendSource(
+                doOnSuccess = { source ->
+                    emit(ViewResource.Success(BannerMapper.toViewParam(source.payload)))
                 },
-                doOnError = {
-                    emit(ViewResource.Error(it.exception))
+                doOnError = { error ->
+                    emit(ViewResource.Error(error.exception))
                 }
             )
         }
